fix ordermathcing tests and eth_with_tokens tests
