+++ mm2src/mm2_main/tests/mm2_tests/orderbook_sync_tests.rs
@@ -6,9 +6,9 @@ use mm2_main::mm2::lp_ordermatch::MIN_ORDER_KEEP_ALIVE_INTERVAL;
 use mm2_number::{BigDecimal, BigRational, MmNumber};
 use mm2_rpc::data::legacy::{AggregatedOrderbookEntry, CoinInitResponse, OrderbookResponse};
 use mm2_test_helpers::electrums::doc_electrums;
-use mm2_test_helpers::for_tests::{enable_z_coin_light, eth_jst_testnet_conf, eth_testnet_conf, get_passphrase,
-                                  morty_conf, orderbook_v2, rick_conf, zombie_conf, MarketMakerIt, Mm2TestConf,
-                                  DOC_ELECTRUM_ADDRS, ETH_DEV_NODES, MARTY_ELECTRUM_ADDRS, RICK, ZOMBIE_ELECTRUMS,
+use mm2_test_helpers::for_tests::{enable_z_coin_light, eth_testnet_conf, get_passphrase, morty_conf, orderbook_v2,
+                                  rick_conf, zombie_conf, MarketMakerIt, Mm2TestConf, DOC_ELECTRUM_ADDRS,
+                                  ETH_DEV_NODES, MARTY_ELECTRUM_ADDRS, RICK, ZOMBIE_ELECTRUMS,
                                   ZOMBIE_LIGHTWALLETD_URLS, ZOMBIE_TICKER};
 use mm2_test_helpers::get_passphrase;
 use mm2_test_helpers::structs::{GetPublicKeyResult, OrderbookV2Response, RpcV2Response, SetPriceResponse};
@@ -1047,7 +1047,7 @@ fn orderbook_should_display_base_rel_volumes() {
 fn orderbook_should_work_without_coins_activation() {
     let bob_passphrase = get_passphrase(&".env.seed", "BOB_PASSPHRASE").unwrap();
 
-    let coins = json!([rick_conf(), morty_conf(), eth_testnet_conf(), eth_jst_testnet_conf(),]);
+    let coins = json!([rick_conf(), morty_conf(), eth_testnet_conf()]);
 
     let mm_bob = MarketMakerIt::start(
         json!({
@@ -1098,7 +1098,7 @@ fn orderbook_should_work_without_coins_activation() {
         "userpass": mm_bob.userpass,
         "method": "setprice",
         "base": "ETH",
-        "rel": "JST",
+        "rel": "RICK",
         "price": "1",
         "volume": "10",
         "min_volume": "1",
@@ -1106,12 +1106,12 @@ fn orderbook_should_work_without_coins_activation() {
     .unwrap();
     assert!(rc.0.is_success(), "!setprice: {}", rc.1);
 
-    log!("Get ETH/JST orderbook on Alice side");
+    log!("Get ETH/RICK orderbook on Alice side");
     let rc = block_on(mm_alice.rpc(&json!({
         "userpass": mm_alice.userpass,
         "method": "orderbook",
         "base": "ETH",
-        "rel": "JST",
+        "rel": "RICK",
     })))
     .unwrap();
     assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
@@ -1119,7 +1119,7 @@ fn orderbook_should_work_without_coins_activation() {
     let orderbook: Json = json::from_str(&rc.1).unwrap();
     log!("orderbook {:?}", orderbook);
     let asks = orderbook["asks"].as_array().unwrap();
-    assert_eq!(asks.len(), 1, "Alice ETH/JST orderbook must have exactly 1 ask");
+    assert_eq!(asks.len(), 1, "Alice ETH/RICK orderbook must have exactly 1 ask");
 }
 
 #[test]
@@ -1199,7 +1199,7 @@ fn test_all_orders_per_pair_per_node_must_be_displayed_in_orderbook() {
 fn setprice_min_volume_should_be_displayed_in_orderbook() {
     let bob_passphrase = get_passphrase(&".env.seed", "BOB_PASSPHRASE").unwrap();
 
-    let coins = json!([rick_conf(), morty_conf(), eth_testnet_conf(), eth_jst_testnet_conf(),]);
+    let coins = json!([rick_conf(), morty_conf(), eth_testnet_conf()]);
 
     let mm_bob = MarketMakerIt::start(
         json!({
@@ -1255,7 +1255,7 @@ fn setprice_min_volume_should_be_displayed_in_orderbook() {
         "userpass": mm_alice.userpass,
         "method": "orderbook",
         "base": "ETH",
-        "rel": "JST",
+        "rel": "RICK",
     })))
     .unwrap();
 
@@ -1263,7 +1263,7 @@ fn setprice_min_volume_should_be_displayed_in_orderbook() {
         "userpass": mm_bob.userpass,
         "method": "setprice",
         "base": "ETH",
-        "rel": "JST",
+        "rel": "RICK",
         "price": "1",
         "volume": "10",
         "min_volume": "1",
@@ -1272,12 +1272,12 @@ fn setprice_min_volume_should_be_displayed_in_orderbook() {
     assert!(rc.0.is_success(), "!setprice: {}", rc.1);
 
     thread::sleep(Duration::from_secs(2));
-    log!("Get ETH/JST orderbook on Bob side");
+    log!("Get ETH/RICK orderbook on Bob side");
     let rc = block_on(mm_bob.rpc(&json!({
         "userpass": mm_bob.userpass,
         "method": "orderbook",
         "base": "ETH",
-        "rel": "JST",
+        "rel": "RICK",
     })))
     .unwrap();
     assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
@@ -1285,17 +1285,17 @@ fn setprice_min_volume_should_be_displayed_in_orderbook() {
     let orderbook: Json = json::from_str(&rc.1).unwrap();
     log!("orderbook {:?}", orderbook);
     let asks = orderbook["asks"].as_array().unwrap();
-    assert_eq!(asks.len(), 1, "Bob ETH/JST orderbook must have exactly 1 ask");
+    assert_eq!(asks.len(), 1, "Bob ETH/RICK orderbook must have exactly 1 ask");
 
     let min_volume = asks[0]["min_volume"].as_str().unwrap();
-    assert_eq!(min_volume, "1", "Bob ETH/JST ask must display correct min_volume");
+    assert_eq!(min_volume, "1", "Bob ETH/RICK ask must display correct min_volume");
 
-    log!("Get ETH/JST orderbook on Alice side");
+    log!("Get ETH/RICK orderbook on Alice side");
     let rc = block_on(mm_alice.rpc(&json!({
         "userpass": mm_alice.userpass,
         "method": "orderbook",
         "base": "ETH",
-        "rel": "JST",
+        "rel": "RICK",
     })))
     .unwrap();
     assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
@@ -1303,10 +1303,10 @@ fn setprice_min_volume_should_be_displayed_in_orderbook() {
     let orderbook: Json = json::from_str(&rc.1).unwrap();
     log!("orderbook {:?}", orderbook);
     let asks = orderbook["asks"].as_array().unwrap();
-    assert_eq!(asks.len(), 1, "Alice ETH/JST orderbook must have exactly 1 ask");
+    assert_eq!(asks.len(), 1, "Alice ETH/RICK orderbook must have exactly 1 ask");
 
     let min_volume = asks[0]["min_volume"].as_str().unwrap();
-    assert_eq!(min_volume, "1", "Alice ETH/JST ask must display correct min_volume");
+    assert_eq!(min_volume, "1", "Alice ETH/RICK ask must display correct min_volume");
 }
 
 // ignored because it requires a long-running ZOMBIE initialization process
